// types/book.ts
export interface Book {
    id: number;
    title: string;
    author: string;
    category: string;
    rating?: number;
    authorId: string;
    categoryId: string;
    description: string;
    price: number;
    coverUrl: string;
    authorName:string;
    categoryName: string;
    categoryname: string;
    averageRating?: number; 

  }
  